Java - TemperatureConverter
In this video I'm going to build a simple temperature converter calculator. 
The purpose is to convert a given temperature in fahrenheit to its equivalent result in celsius. 
As tools to develop this program, we use printf and Scanner

You can check the video where I type this code at:
https://www.youtube.com/watch?v=nhrPfwoxOWE